ANIMATION: 2D mode
Turns out, it really wasn't the vertex matrix multiplication that made
the 3D animation so intensive. It really is just the MatLab plot tool
overhead. So despite the newly implemented 2D animation mode being
extremely simplistic, its use of the plot tool still makes it a bit
slow. However, it is a bit easier on the computer, so it serves its
purpose... a little. Regardless, it's here now.

Also, now if the boat is out of the water, the thrusters have no effect.